Further refactoring for extensions
